realization battle

+++ WarriorLibrary/SuperPowerGenerator.cs
@@ -3,8 +3,8 @@
 public static class SuperPowerGenerator
 {
     private static int _counter;
-    public static int CapacitySuperPower(int lowerNumber = 5, int upperNumber = 10)
+    public static uint CapacitySuperPower(uint lowerNumber = 5, uint upperNumber = 10)
     {
-        return (DateTime.Now.Millisecond + _counter++) % (upperNumber-lowerNumber+1) + lowerNumber;
+        return (uint) ((DateTime.Now.Millisecond + _counter++) % (upperNumber-lowerNumber+1) + lowerNumber);
     }
 }
